+++ src/main/java/pl/betoncraft/betonquest/events/TagEvent.java
@@ -17,14 +17,21 @@
  */
 package pl.betoncraft.betonquest.events;
 
+import org.bukkit.Bukkit;
+import org.bukkit.entity.Player;
 import org.bukkit.scheduler.BukkitRunnable;
 import pl.betoncraft.betonquest.BetonQuest;
 import pl.betoncraft.betonquest.Instruction;
 import pl.betoncraft.betonquest.api.QuestEvent;
+import pl.betoncraft.betonquest.database.Connector;
 import pl.betoncraft.betonquest.database.PlayerData;
+import pl.betoncraft.betonquest.database.Saver;
 import pl.betoncraft.betonquest.exceptions.InstructionParseException;
+import pl.betoncraft.betonquest.utils.LogUtils;
 import pl.betoncraft.betonquest.utils.PlayerConverter;
 import pl.betoncraft.betonquest.utils.Utils;
+
+import java.util.logging.Level;
 
 /**
  * Adds or removes tags from the player
@@ -39,6 +46,7 @@ public class TagEvent extends QuestEvent {
     public TagEvent(Instruction instruction) throws InstructionParseException {
         super(instruction, false);
         persistent = true;
+        staticness = true;
         add = instruction.next().equalsIgnoreCase("add");
         tags = instruction.getArray();
         for (int i = 0; i < tags.length; i++) {
@@ -48,7 +56,19 @@ public class TagEvent extends QuestEvent {
 
     @Override
     protected Void execute(final String playerID) {
-        if (PlayerConverter.getPlayer(playerID) != null) {
+        if (playerID == null) {
+            if (!add) {
+                for (String tag : tags) {
+                    for (Player p : Bukkit.getOnlinePlayers()) {
+                        PlayerData playerData = BetonQuest.getInstance().getPlayerData(PlayerConverter.getID(p));
+                        playerData.removeTag(tag);
+                    }
+                    BetonQuest.getInstance().getSaver().add(new Saver.Record(Connector.UpdateType.REMOVE_ALL_TAGS, new String[]{
+                            tag
+                    }));
+                }
+            }
+        } else if (PlayerConverter.getPlayer(playerID) != null) {
             PlayerData playerData = BetonQuest.getInstance().getPlayerData(playerID);
             if (add) {
                 for (String tag : tags) {